pyproject.toml
[tool.poetry]
name = "terra-sdk"
authors = ["Terraform Labs, PTE. <engineering@terra.money>"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10"
]
description = "The Python SDK for Terra"
documentation = "https://terra-money.github.io/terra.py/index.html"
homepage = "https://github.com/terra-money/terra.py"
keywords = ["jigu", "blockchain", "terra", "defi", "crypto"]
license = "MIT"
packages = [{ include = "terra_sdk" }]
readme = "README.md"
repository = "https://github.com/terra-money/terra.py"
version = "3.0.0"

[tool.poetry.dependencies]
aiohttp = "^3.8.1"
bech32 = "^1.2.0"
bip32utils = "^0.3.post4"
ecdsa = "^0.17.0"
mnemonic = "^0.19"
python = "^3.7"
protobuf = "^3.19.1"
betterproto = "2.0.0b4"
furl = "^2.1.3"
boltons = "^21.0.0"
nest-asyncio = "^1.5.4"
attrs = "^21.4.0"
wrapt = "^1.13.3"
terra-proto = "1.1.0"

[tool.poetry.dev-dependencies]
aioresponses = "^0.7.2"
asynctest = "^0.13.0"
pytest-cov = "^3.0.0"
pytest = "^7.0.1"
pip = "^22.0.3"
Sphinx = "^4.4.0"
sphinx-autobuild = "^2021.3.14"
black = "^22.1.0"
coverage = "^6.3.2"
isort = "^5.10.1"
mypy = "^0.931"
pytest-mock = "^3.7.0"
pytest-runner = "^5.3.1"
pytest-sugar = "^0.9.4"
sphinx-autodoc-typehints = "^1.17.0"
uvloop = "^0.16.0"
watchdog = "^2.1.6"
wheel = "^0.37.1"
flake8 = "^3.9.2"
sphinx-rtd-theme = "^1.0.0"
types-python-dateutil = "^2.8.9"

[tool.pytest.ini_options]
addopts = "-ra -q"
minversion = "6.0"
asyncio_mode = "auto"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
